modularised server file
fix #38
