+++ src/server.js
@@ -7,6 +7,8 @@ const fs = require('fs');
 const path = require('path');
 const haj = require('hapi-auth-jwt2');
 const users = require('../database/users');
+const strategyOptions = require('./helpers')
+
 
 
 const routes = require('./routes');
@@ -25,19 +27,7 @@ server.connection({
 server.register([inert, vision, cookieAuth, haj], (err) => {
   if (err) throw err;
 
-  const validate = (token, request, callback) => {
-    console.log("token", token); // decoded token, it automatically decodes it
-    if (!users[token.user.user_id]) {
-      return callback(null, false);
-    }
-    return callback(null, true);
-  }
 
-  const strategyOptions = {
-    key: process.env.JWT_SECRET,
-    validateFunc: validate,
-    verifyOptions: { algorithms: [ 'HS256' ] }
-  }
 
   server.auth.strategy('jwt', 'jwt', strategyOptions);
